Allow "adb shell tradeinmode" on user builds.
We initially wanted "adb shell tradeinmode" to work on userdebug/eng for
CI testing. But, we need xTS tests as well. These run in a special
ro.debuggable user build, so we need to adjust our sepolicy.

As a result of this change, the trade-in mode binary now checks
ro.debuggable to forbid unintended use. That CL is ag/30497218.

Bug: 379970773
Test: adb shell tradeinmode on cf-user
Change-Id: If23f4dc02be32a3aa9db2cf1848b499bfc3f71d6