Add permission to installd to delete `virtualizationservice_data_file`
Terminal app had re-labelled its data file to `virtualizationservice_data_file`,
but we don't use that label anymore, so removed permission as well.

But it causes sepolicy error when updating from older version.
At least to recover with clean image, we need to provide a way to clear
terminal app in developer's option.

Bug: 383026786
Change-Id: I4563156c436e34c0adbf81799e709391fde6d0b4
Test: install older version and update terminal app, while preserving data