Allow ashmem ioctl commands to be invoked on tmpfs files
Part of the migration from ashmem to memfd involves adding support for
ashmem ioctl commands to memfds. This allows applications that directly
interact with ashmem buffers via ioctls to continue working seamlessly
even if they are given a memfd buffer instead of an ashmem buffer to
operate on.

Therefore, allow ioctl() to be invoked on tmpfs files.

Bug: 111903542
Change-Id: I68c269af54764e1abc6346e05bbf94acb99c44f1
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>