Allow ephemeral apps to access the protolog configuration service
Ephemeral like normal apps write to protolog which needs access to the protolog configuration service to determine which tracing configuration to use and keep the configuration in sync across all apps.

This lack of permission provided to ephemeral apps is currently causing IME CTS tests which use instant apps to fail.

Bug: 374157530
Change-Id: I84a42f9038fa4e39fa18adf268420e5be5f563e3